// frontend/lib/supabase/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5";
  };
  public: {
    Tables: {
      company: {
        Row: {
          average_age: number | null;
          average_salary: number | null;
          benefits: string | null;
          contact_name: string | null;
          created_at: string;
          description: string | null;
          id: string;
          industry: string | null;
          logo_url: string | null;
          name: string;
          seeking_people: string | null;
          strengths: string | null;
          updated_at: string | null;
          wallet_address: string | null;
        };
        Insert: {
          average_age?: number | null;
          average_salary?: number | null;
          benefits?: string | null;
          contact_name?: string | null;
          created_at?: string;
          description?: string | null;
          id?: string;
          industry?: string | null;
          logo_url?: string | null;
          name: string;
          seeking_people?: string | null;
          strengths?: string | null;
          updated_at?: string | null;
          wallet_address?: string | null;
        };
        Update: {
          average_age?: number | null;
          average_salary?: number | null;
          benefits?: string | null;
          contact_name?: string | null;
          created_at?: string;
          description?: string | null;
          id?: string;
          industry?: string | null;
          logo_url?: string | null;
          name?: string;
          seeking_people?: string | null;
          strengths?: string | null;
          updated_at?: string | null;
          wallet_address?: string | null;
        };
        Relationships: [];
      };
      company_link_user: {
        Row: {
          company_id: string | null;
          created_at: string;
          id: number;
          role: number | null;
          user_id: string | null;
        };
        Insert: {
          company_id?: string | null;
          created_at?: string;
          id?: number;
          role?: number | null;
          user_id?: string | null;
        };
        Update: {
          company_id?: string | null;
          created_at?: string;
          id?: number;
          role?: number | null;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "company_link_user_company_id_fkey";
            columns: ["company_id"];
            isOneToOne: false;
            referencedRelation: "company";
            referencedColumns: ["id"];
          },
        ];
      };
      student: {
        Row: {
          about: string | null;
          created_at: string;
          department: string | null;
          faculty: string | null;
          grade: number | null;
          icon_url: string | null;
          id: string;
          name: string;
          skill: string | null;
          univercity: string | null;
          updated_at: string | null;
          wallet_address: string | null;
        };
        Insert: {
          about?: string | null;
          created_at?: string;
          department?: string | null;
          faculty?: string | null;
          grade?: number | null;
          icon_url?: string | null;
          id?: string;
          name: string;
          skill?: string | null;
          univercity?: string | null;
          updated_at?: string | null;
          wallet_address?: string | null;
        };
        Update: {
          about?: string | null;
          created_at?: string;
          department?: string | null;
          faculty?: string | null;
          grade?: number | null;
          icon_url?: string | null;
          id?: string;
          name?: string;
          skill?: string | null;
          univercity?: string | null;
          updated_at?: string | null;
          wallet_address?: string | null;
        };
        Relationships: [];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema =
  DatabaseWithoutInternals[Extract<keyof Database, "public">];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]][
      "Enums"
    ]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][
    EnumName
  ]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
  : never;

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[
      PublicCompositeTypeNameOrOptions["schema"]
    ]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]][
    "CompositeTypes"
  ][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends